// web-application/src/types-openapi/api/models/CreateUserMembershipResponse.ts
export interface CreateUserMembershipResponse {
    /**
     * 
     * @type {number}
     * @memberof CreateUserMembershipResponse
     */
    userMembershipId?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateUserMembershipResponse
     */
    customerId?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateUserMembershipResponse
     */
    membershipId?: number;
    /**
     * 
     * @type {string}
     * @memberof CreateUserMembershipResponse
     */
    status?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CreateUserMembershipResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof CreateUserMembershipResponse
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CreateUserMembershipResponse
     */
    endDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof CreateUserMembershipResponse
     */
    paymentId?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreateUserMembershipResponse
     */
    paymentAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof CreateUserMembershipResponse
     */
    paymentMethod?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateUserMembershipResponse
     */
    qrUrl?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreateUserMembershipResponse
     */
    holdMinutes?: number;
    /**
     * 
     * @type {Date}
     * @memberof CreateUserMembershipResponse
     */
    expiresAtUtc?: Date;
}